+++ stories/Components/accordion.stories.tsx
@@ -5,31 +5,7 @@ import { storiesOf } from '@storybook/react';
 import { GlobalStyles } from '@atlantum/theme';
 import { Card } from '@atlantum/card';
 import { NavigationLink } from '@atlantum/navigation-link';
-import { ItemsStack } from '@atlantum/items-stack';
 import '../stories.styles.css';
-
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'row'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A507"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A502"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
 
 const Accordions = () => {
     return (
@@ -77,7 +53,6 @@ const Accordions = () => {
     );
 };
 
-storiesOf('Components/Accordion', module).add('Design', () => <Design />);
 storiesOf('Components/Accordion', module).add('Accordions', () => (
     <Accordions />
 ));