+++ stories/Components/alert.stories.tsx
@@ -8,29 +8,6 @@ import { Card } from '@atlantum/card';
 import { Heading } from '@atlantum/heading';
 import '../stories.styles.css';
 
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'column'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A321"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ border: '1px solid rgba(0, 0, 0, 0.1)' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A2"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
-
 const AlertInfo = () => {
     return (
         <>
@@ -131,7 +108,6 @@ const AlertDanger = () => {
     );
 };
 
-storiesOf('Components/Alert', module).add('Design', () => <Design />);
 storiesOf('Components/Alert', module).add('Alert info', () => <AlertInfo />);
 storiesOf('Components/Alert', module).add('Alert success', () => (
     <AlertSuccess />