+++ stories/Components/buttons.stories.tsx
@@ -9,29 +9,6 @@ import '../stories.styles.css';
 import { Card } from '@atlantum/card';
 import { ItemsStack } from '@atlantum/items-stack';
 
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'column'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A3192"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A3183"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
-
 const Primary = () => {
     return (
         <>
@@ -135,7 +112,6 @@ const BtnGroup = () => {
     );
 };
 
-storiesOf('Components/Button', module).add('Design', () => <Design />);
 storiesOf('Components/Button', module).add('Primary', () => <Primary />);
 storiesOf('Components/Button', module).add('Secondary', () => <Secondary />);
 storiesOf('Components/Button', module).add('Disabled', () => <Disabled />);