+++ stories/Components/datepicker.stories.tsx
@@ -1,20 +1,9 @@
 import * as React from 'react';
-import { Datepicker } from '../../packages/datepicker';
+import { Datepicker } from '@atlantum/datepicker';
 import { Card } from '@atlantum/card';
 import { storiesOf } from '@storybook/react';
 import { GlobalStyles } from '@atlantum/theme';
 import '../stories.styles.css';
-
-const Design = () => {
-    return (
-        <>
-            <GlobalStyles />
-            <div className="atlantum-compact">
-                <Card size="standard">Under development</Card>
-            </div>
-        </>
-    );
-};
 
 const DefaultInput = () => {
     return (
@@ -35,7 +24,6 @@ const DefaultInput = () => {
     );
 };
 
-storiesOf('Components/Datepicker', module).add('Design', () => <Design />);
 storiesOf('Components/Datepicker', module).add('Default', () => (
     <DefaultInput />
 ));