+++ stories/Components/divider.stories.tsx
@@ -7,29 +7,6 @@ import { Divider } from '@atlantum/divider';
 
 import '../stories.styles.css';
 
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'row'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D811%253A3"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ border: '1px solid rgba(0, 0, 0, 0.1)' }}
-                    width="100%"
-                    height="300"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D811%253A0"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
-
 const DividerDefault = () => {
     return (
         <>
@@ -46,7 +23,6 @@ const DividerDefault = () => {
     );
 };
 
-storiesOf('Components/Divider', module).add('Design', () => <Design />);
 storiesOf('Components/Divider', module).add('Default', () => (
     <DividerDefault />
 ));