+++ stories/Components/table.stories.tsx
@@ -2,34 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { GlobalStyles } from '@atlantum/theme';
 import { Card } from '@atlantum/card';
-import { ItemsStack } from '@atlantum/items-stack';
+
 import { Table, TableCell, TableHeader, TableRow } from '@atlantum/table';
 import '../stories.styles.css';
-
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'row'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D127%253A768"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D127%253A763"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
-
-storiesOf('Components/Table', module).add('Design', () => <Design />);
 
 const Basic = () => {
     return (