+++ stories/Fundamentals/typography.stories.tsx
@@ -7,20 +7,6 @@ import { Text } from '@atlantum/text';
 import { Card } from '@atlantum/card';
 import { Group } from '@atlantum/group';
 
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <iframe
-                style={{ borderRadius: '4px', border: 'none' }}
-                width="100%"
-                height="600"
-                src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D344%253A1368"
-                allowFullScreen
-            />
-        </div>
-    );
-};
-
 const Headlines = () => {
     return (
         <>
@@ -66,8 +52,6 @@ const Paragraph = () => {
         </>
     );
 };
-
-storiesOf('Fundamentals/Typography', module).add('Design', () => <Design />);
 
 storiesOf('Fundamentals/Typography', module).add('Heading', () => (
     <Headlines />