+++ stories/Layout/card.stories.tsx
@@ -6,30 +6,6 @@ import { Card } from '@atlantum/card';
 import { Box } from '../utilities/box.ut';
 
 import '../stories.styles.css';
-import { ItemsStack } from '@atlantum/items-stack';
-
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction={'row'}>
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A3264"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D96%253A3255"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
 
 const Full = () => {
     return (
@@ -96,7 +72,6 @@ const Landing = () => {
     );
 };
 
-storiesOf('Layout/Card', module).add('Design', () => <Design />);
 storiesOf('Layout/Card', module).add('Full', () => <Full />);
 storiesOf('Layout/Card', module).add('Large', () => <Large />);
 storiesOf('Layout/Card', module).add('Standard', () => <Standard />);