+++ stories/Layout/group.stories.tsx
@@ -10,30 +10,6 @@ import { Group } from '@atlantum/group';
 import '../stories.styles.css';
 
 import { Box } from '../utilities/box.ut';
-import { ItemsStack } from '@atlantum/items-stack';
-
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <ItemsStack direction="row">
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D189%253A123"
-                    allowFullScreen
-                />
-                <iframe
-                    style={{ borderRadius: '4px', border: 'none' }}
-                    width="100%"
-                    height="400"
-                    src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D189%253A121"
-                    allowFullScreen
-                />
-            </ItemsStack>
-        </div>
-    );
-};
 
 const Responsive = () => {
     return (