+++ stories/Layout/items-stack.stories.tsx
@@ -8,29 +8,6 @@ import { ItemsStack } from '@atlantum/items-stack';
 import { Text } from '@atlantum/text';
 import { Card } from '@atlantum/card';
 
-const Design = () => {
-    return (
-        <div className="atlantum-full-view">
-            <iframe
-                style={{ borderRadius: '4px', border: 'none' }}
-                width="100%"
-                height="400"
-                src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D189%253A123"
-                allowFullScreen
-            />
-            <iframe
-                style={{ borderRadius: '4px', border: 'none' }}
-                width="100%"
-                height="400"
-                src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2F5CdA4Nbd80vV0fy43Or0kZ%2FAtlantum-Examples%3Fnode-id%3D189%253A121"
-                allowFullScreen
-            />
-        </div>
-    );
-};
-
-storiesOf('Layout/ItemsStack', module).add('Design', () => <Design />);
-
 const Horizontal = () => {
     return (
         <>