+++ packages/tag/CHANGELOG.md
@@ -1,6 +1,7 @@
 # @atlantum/tag
 
 ## 0.0.2
+
 ### Patch Changes
 
-- 174727b: initialized tag design
+-   174727b: initialized tag design